Fix subtle issue where wrong version was being compiled against (0.20.1).
This lead to a compile error in the IDE of the library project. The demo app compiles fine because it relies on its own version of react native but the library project could not compile, independently. This suggests that the react native version for the overall project should be managed in one place but the current configuration is probably sufficient for a proof of concept app.