Upgrade SDK
The block processor fires many notifications on initial start up but becomes very quiet once synced. Once synced we only need to update the processor state when the blockheight increases. We can listen to all events but only pass along to the consumer when one of the watched variabled changes